requirements.txt
alembic==1.11.3
blinker==1.6.2
cffi==1.15.1
click==8.1.7
cryptography==41.0.3
dnspython==2.4.2
dominate==2.8.0
email-validator==2.0.0.post2
Flask==2.3.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
visitor==0.1.3
Werkzeug==2.3.7
WTForms==3.0.1
